requirements.txt
pdfplumber>=0.10.0
gspread>=5.12.0
google-auth>=2.23.0
PyYAML>=6.0.1

